requirements.txt
openai==2.28.0
python-dotenv==1.2.2
pydantic==2.12.5
